# copilot-network-analyzer/connector/apiDefinition.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ntopng Network Analyzer",
    "description": "Read-only network visibility over ntopng: active hosts, live flows, protocol breakdown and alerts. Built to back a Microsoft Copilot Studio agent, but usable by any HTTP client.",
    "version": "0.1.0"
  },
  "host": "REPLACE-WITH-YOUR-PUBLIC-HOSTNAME",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/hosts/top": {
      "get": {
        "operationId": "GetTopTalkers",
        "summary": "Get the busiest hosts on the network right now",
        "description": "Returns the hosts currently generating the most traffic, ordered by live throughput or by total bytes. Use this to answer questions like 'what is using all the bandwidth', 'who are the top talkers', or 'what is the busiest device right now'.",
        "x-ms-visibility": "important",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "How many hosts to return.",
            "x-ms-summary": "Limit",
            "type": "integer",
            "default": 10,
            "minimum": 1,
            "maximum": 100
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "description": "Rank by live throughput or by cumulative bytes.",
            "x-ms-summary": "Order By",
            "type": "string",
            "default": "throughput",
            "pattern": "^(throughput|bytes)$"
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "description": "Limit to local hosts, remote hosts, or return both.",
            "x-ms-summary": "Scope",
            "type": "string",
            "default": "all",
            "pattern": "^(all|local|remote)$"
          },
          {
            "name": "interface_id",
            "in": "query",
            "required": false,
            "description": "ntopng interface id. Defaults to the configured one.",
            "x-ms-summary": "Interface Id",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/HostList"
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "502": {
            "description": "ntopng returned an error."
          },
          "503": {
            "description": "ntopng is unreachable."
          }
        }
      }
    },
    "/v1/hosts/search": {
      "get": {
        "operationId": "FindHost",
        "summary": "Find an active host by name or IP address",
        "description": "Case-insensitive substring search over the names and IP addresses of currently active hosts. Use this when the user names a device, for example 'is the living room TV online' or 'show me 10.9.1.50'.",
        "x-ms-visibility": "important",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "description": "Part of a host name or IP address, e.g. 'iphone' or '10.9.1.'.",
            "x-ms-summary": "Query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "How many matches to return.",
            "x-ms-summary": "Limit",
            "type": "integer",
            "default": 10,
            "minimum": 1,
            "maximum": 100
          },
          {
            "name": "interface_id",
            "in": "query",
            "required": false,
            "description": "ntopng interface id. Defaults to the configured one.",
            "x-ms-summary": "Interface Id",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/HostList"
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "502": {
            "description": "ntopng returned an error."
          },
          "503": {
            "description": "ntopng is unreachable."
          }
        }
      }
    },
    "/v1/flows/active": {
      "get": {
        "operationId": "GetActiveFlows",
        "summary": "List the active network conversations",
        "description": "Returns currently active flows (client, server, application protocol and volume), heaviest first, optionally filtered to one host or one protocol. Use this to answer 'who is my laptop talking to', 'what is this device downloading', or 'show me current connections'.",
        "x-ms-visibility": "important",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "How many flows to return.",
            "x-ms-summary": "Limit",
            "type": "integer",
            "default": 20,
            "minimum": 1,
            "maximum": 200
          },
          {
            "name": "host",
            "in": "query",
            "required": false,
            "description": "Restrict to flows involving this IP address or host name.",
            "x-ms-summary": "Host",
            "type": "string"
          },
          {
            "name": "l7_protocol",
            "in": "query",
            "required": false,
            "description": "Restrict to one application protocol, e.g. HTTPS, DNS, Netflix.",
            "x-ms-summary": "L7 Protocol",
            "type": "string"
          },
          {
            "name": "interface_id",
            "in": "query",
            "required": false,
            "description": "ntopng interface id. Defaults to the configured one.",
            "x-ms-summary": "Interface Id",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/FlowList"
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "502": {
            "description": "ntopng returned an error."
          },
          "503": {
            "description": "ntopng is unreachable."
          }
        }
      }
    },
    "/v1/protocols/top": {
      "get": {
        "operationId": "GetTopProtocols",
        "summary": "Break traffic down by application protocol",
        "description": "Returns the heaviest application protocols detected by nDPI, with each one's share of total traffic. Use this for 'what kind of traffic is on my network', 'how much is streaming', or 'what protocol is dominant'.",
        "x-ms-visibility": "important",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "How many protocols to return.",
            "x-ms-summary": "Limit",
            "type": "integer",
            "default": 10,
            "minimum": 1,
            "maximum": 50
          },
          {
            "name": "interface_id",
            "in": "query",
            "required": false,
            "description": "ntopng interface id. Defaults to the configured one.",
            "x-ms-summary": "Interface Id",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/ProtocolList"
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "502": {
            "description": "ntopng returned an error."
          },
          "503": {
            "description": "ntopng is unreachable."
          }
        }
      }
    }
  },
  "definitions": {
...
    "Flow": {
      "properties": {
        "client_ip": {
          "type": "string",
          "title": "Client Ip",
          "description": "IP address of the client side of the flow."
        },
        "client_name": {
          "type": "string",
          "title": "Client Name",
          "description": "Resolved name of the client, when known.",
          "x-nullable": true
        },
        "server_ip": {
          "type": "string",
          "title": "Server Ip",
          "description": "IP address of the server side of the flow."
        },
        "server_name": {
          "type": "string",
          "title": "Server Name",
          "description": "Resolved name of the server, when known.",
          "x-nullable": true
        },
        "server_port": {
          "type": "integer",
          "title": "Server Port",
          "description": "Destination (server) port.",
          "x-nullable": true
        },
        "l7_protocol": {
          "type": "string",
          "title": "L7 Protocol",
          "description": "Application protocol detected by nDPI, e.g. HTTPS, DNS, Netflix.",
          "x-nullable": true
        },
        "l4_protocol": {
          "type": "string",
          "title": "L4 Protocol",
          "description": "Transport protocol, e.g. TCP or UDP.",
          "x-nullable": true
        },
        "bytes_total": {
          "type": "integer",
          "title": "Bytes Total",
          "description": "Bytes exchanged on this flow so far."
        },
        "throughput_bps": {
          "type": "number",
          "title": "Throughput Bps",
          "description": "Current throughput of the flow in bits per second.",
          "default": 0
        }
      },
      "type": "object",
      "required": [
        "client_ip",
        "server_ip",
        "bytes_total"
      ],
      "title": "Flow"
    },
    "FlowList": {
      "properties": {
        "summary": {
          "type": "string",
          "title": "Summary",
          "description": "One-sentence, human-readable summary of the result."
        },
        "interface_id": {
          "type": "integer",
          "title": "Interface Id",
          "description": "ntopng interface the data came from."
        },
        "count": {
          "type": "integer",
          "title": "Count",
          "description": "Number of flows returned."
        },
        "flows": {
          "items": {
            "$ref": "#/definitions/Flow"
          },
          "type": "array",
          "title": "Flows",
          "description": "The active flows, largest first."
        }
      },
      "type": "object",
      "required": [
        "summary",
        "interface_id",
        "count",
        "flows"
      ],
      "title": "FlowList"
    },
...
    "Host": {
      "properties": {
        "ip": {
          "type": "string",
          "title": "Ip",
          "description": "IP address of the host."
        },
        "name": {
          "type": "string",
          "title": "Name",
          "description": "Resolved name (mDNS/DHCP/reverse DNS), or the IP when unresolved."
        },
        "is_local": {
          "type": "boolean",
          "title": "Is Local",
          "description": "True when the host is inside a configured local network."
        },
        "country": {
          "type": "string",
          "title": "Country",
          "description": "Two-letter country code for remote hosts.",
          "x-nullable": true
        },
        "throughput_bps": {
          "type": "number",
          "title": "Throughput Bps",
          "description": "Current throughput in bits per second."
        },
        "bytes_total": {
          "type": "integer",
          "title": "Bytes Total",
          "description": "Total bytes seen for this host since ntopng started tracking it."
        },
        "bytes_sent": {
          "type": "integer",
          "title": "Bytes Sent",
          "description": "Bytes sent by the host."
        },
        "bytes_received": {
          "type": "integer",
          "title": "Bytes Received",
          "description": "Bytes received by the host."
        },
        "active_flows": {
          "type": "integer",
          "title": "Active Flows",
          "description": "Number of currently active flows involving the host."
        },
        "last_seen": {
          "type": "integer",
          "title": "Last Seen",
          "description": "Unix timestamp of the last packet seen.",
          "x-nullable": true
        }
      },
      "type": "object",
      "required": [
        "ip",
        "name",
        "is_local",
        "throughput_bps",
        "bytes_total",
        "bytes_sent",
        "bytes_received",
        "active_flows"
      ],
      "title": "Host"
    },
    "HostList": {
      "properties": {
        "summary": {
          "type": "string",
          "title": "Summary",
          "description": "One-sentence, human-readable summary of the result."
        },
        "interface_id": {
          "type": "integer",
          "title": "Interface Id",
          "description": "ntopng interface the data came from."
        },
        "count": {
          "type": "integer",
          "title": "Count",
          "description": "Number of hosts returned."
        },
        "hosts": {
          "items": {
            "$ref": "#/definitions/Host"
          },
          "type": "array",
          "title": "Hosts",
          "description": "The hosts, ordered as requested."
        }
      },
      "type": "object",
      "required": [
        "summary",
        "interface_id",
        "count",
        "hosts"
      ],
      "title": "HostList"
    },
...
    "Protocol": {
      "properties": {
        "name": {
          "type": "string",
          "title": "Name",
          "description": "Application protocol name as detected by nDPI."
        },
        "bytes_total": {
          "type": "integer",
          "title": "Bytes Total",
          "description": "Bytes attributed to this protocol."
        },
        "share_percent": {
          "type": "number",
          "title": "Share Percent",
          "description": "Share of total traffic, in percent."
        }
      },
      "type": "object",
      "required": [
        "name",
        "bytes_total",
        "share_percent"
      ],
      "title": "Protocol"
    },
    "ProtocolList": {
      "properties": {
        "summary": {
          "type": "string",
          "title": "Summary",
          "description": "One-sentence, human-readable summary of the result."
        },
        "interface_id": {
          "type": "integer",
          "title": "Interface Id",
          "description": "ntopng interface the data came from."
        },
        "protocols": {
          "items": {
            "$ref": "#/definitions/Protocol"
          },
          "type": "array",
          "title": "Protocols",
          "description": "Protocols, heaviest first."
        }
      },
      "type": "object",
      "required": [
        "summary",
        "interface_id",
        "protocols"
      ],
      "title": "ProtocolList"
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-Key"
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}